begin manage folders popup

+++ src/Glimpse.API/Database/Track.cs
@@ -21,7 +21,7 @@ public record struct Track
 
     public byte Rating;
 
-    public byte PlayCount;
+    public uint PlayCount;
 
     public DateTime? LastPlayed;
